🧼 Renamed object
WeatherAPIResponse was not an appropriate name, as the object was modelling the weather itself. In addition, WeatherAPIResponse needs to be used for at action object for the respective endpoint.
